add free_bfs.c
