+++ srcs/bfs_free.c
@@ -1,0 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   bfs_free.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: thule <thule@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/10/10 11:15:03 by thle              #+#    #+#             */
+/*   Updated: 2022/11/02 11:44:17 by itkimura         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/lemin.h"
+
+void free_path(t_path *path)
+{
+	free(path->path);
+	free(path);
+	path = NULL;
+}
+
+void free_que(t_que *head)
+{
+	t_que *next;
+
+	while (head)
+	{
+		next = head->next;
+		free(head);
+		head = next;
+	}
+}
